+++ crates/rspack_core/src/dependency/css/import.rs
@@ -51,8 +51,8 @@ impl CssImportDependency {
 }
 
 impl Dependency for CssImportDependency {
-  fn id(&self) -> Option<&DependencyId> {
-    self.id.as_ref()
+  fn id(&self) -> Option<DependencyId> {
+    self.id
   }
   fn set_id(&mut self, id: Option<DependencyId>) {
     self.id = id;