+++ crates/rspack_core/src/dependency/entry.rs
@@ -28,8 +28,8 @@ impl Dependency for EntryDependency {
     &DependencyType::Entry
   }
 
-  fn id(&self) -> Option<&DependencyId> {
-    self.id.as_ref()
+  fn id(&self) -> Option<DependencyId> {
+    self.id
   }
 
   fn set_id(&mut self, id: Option<DependencyId>) {