+++ crates/rspack_plugin_runtime/src/runtime_module/require_js_chunk_loading.rs
@@ -3,13 +3,16 @@ use rspack_core::{
   rspack_sources::{BoxSource, ConcatSource, RawSource, SourceExt},
   runtime_globals, ChunkUkey, Compilation, RuntimeModule, SourceType, RUNTIME_MODULE_STAGE_ATTACH,
 };
+use rspack_identifier::Identifier;
 use rustc_hash::FxHashSet as HashSet;
 
 use super::utils::{chunk_has_js, get_undo_path};
+use crate::impl_runtime_module;
 use crate::runtime_module::utils::{get_initial_chunk_ids, stringify_chunks};
 
-#[derive(Debug, Default)]
+#[derive(Debug, Default, Eq)]
 pub struct RequireChunkLoadingRuntimeModule {
+  id: Identifier,
   chunk: Option<ChunkUkey>,
   runtime_requirements: HashSet<&'static str>,
 }
@@ -17,6 +20,7 @@ pub struct RequireChunkLoadingRuntimeModule {
 impl RequireChunkLoadingRuntimeModule {
   pub fn new(runtime_requirements: HashSet<&'static str>) -> Self {
     Self {
+      id: Identifier::from("webpack/runtime/require_chunk_loading"),
       chunk: None,
       runtime_requirements,
     }
@@ -24,8 +28,8 @@ impl RequireChunkLoadingRuntimeModule {
 }
 
 impl RuntimeModule for RequireChunkLoadingRuntimeModule {
-  fn identifier(&self) -> String {
-    "webpack/runtime/require_chunk_loading".to_string()
+  fn name(&self) -> Identifier {
+    self.id
   }
 
   fn generate(&self, compilation: &Compilation) -> BoxSource {
@@ -89,9 +93,9 @@ impl RuntimeModule for RequireChunkLoadingRuntimeModule {
     }
 
     if with_hmr {
-      source.add(RawSource::from(
-        include_str!("runtime/require_chunk_loading_with_hmr.js").to_string(),
-      ));
+      source.add(RawSource::from(include_str!(
+        "runtime/require_chunk_loading_with_hmr.js"
+      )));
       source.add(RawSource::from(
         include_str!("runtime/javascript_hot_module_replacement.js").replace("$key$", "jsonp"),
       ));
@@ -101,24 +105,24 @@ impl RuntimeModule for RequireChunkLoadingRuntimeModule {
       .runtime_requirements
       .contains(runtime_globals::HMR_DOWNLOAD_MANIFEST)
     {
-      source.add(RawSource::from(
-        include_str!("runtime/require_chunk_loading_with_hmr_manifest.js").to_string(),
-      ));
+      source.add(RawSource::from(include_str!(
+        "runtime/require_chunk_loading_with_hmr_manifest.js"
+      )));
     }
 
     if self
       .runtime_requirements
       .contains(runtime_globals::ON_CHUNKS_LOADED)
     {
-      source.add(RawSource::from(
-        include_str!("runtime/require_chunk_loading_with_on_chunk_load.js").to_string(),
-      ));
+      source.add(RawSource::from(include_str!(
+        "runtime/require_chunk_loading_with_on_chunk_load.js"
+      )));
     }
 
     if with_external_install_chunk {
-      source.add(RawSource::from(
-        include_str!("runtime/require_chunk_loading_with_external_install_chunk.js").to_string(),
-      ));
+      source.add(RawSource::from(include_str!(
+        "runtime/require_chunk_loading_with_external_install_chunk.js"
+      )));
     }
 
     source.boxed()
@@ -132,3 +136,5 @@ impl RuntimeModule for RequireChunkLoadingRuntimeModule {
     RUNTIME_MODULE_STAGE_ATTACH
   }
 }
+
+impl_runtime_module!(RequireChunkLoadingRuntimeModule);