+++ examples/vue/rspack.config.js
@@ -13,7 +13,7 @@ module.exports = {
     }
   },
   devServer: {
-    historyApiFallback:true
+    historyApiFallback: true
   },
   module: {
     rules: [