+++ packages/rspack/tests/configCases/asset/issue#2283/index.js
@@ -1,0 +1,9 @@
+import fs from "fs";
+require("./img.png");
+
+it("should create the same asset as input", () => {
+	const png = fs.readdirSync(__dirname).find(file => file.endsWith(".png"));
+	expect(fs.readFileSync(__dirname + "/" + png)).toEqual(
+		fs.readFileSync(__dirname + "/../" + "img.png")
+	);
+});