+++ benchcases/react-refresh/CHANGELOG.md
@@ -1,5 +1,17 @@
 # benchmarkcase-rspack-react-refresh
 
+## 1.0.3
+
+### Patch Changes
+
+- a004765a: Avoid reporting warning the export when src module format is none esm
+- Updated dependencies [7b3167ca]
+- Updated dependencies [c4c20d66]
+- Updated dependencies [a004765a]
+- Updated dependencies [0909f15f]
+- Updated dependencies [3d459261]
+  - @rspack/cli@0.1.1
+
 ## 1.0.2
 
 ### Patch Changes